backend/requirements.txt
# Core Framework
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis[hiredis]==5.0.1
hiredis==2.3.2

# Azure SDKs
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-servicebus==7.11.4

# AI/ML
openai==1.10.0
tiktoken==0.5.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Docker SDK
docker==7.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3

# Validation
email-validator==2.1.0
phonenumbers==8.13.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
